SDL/smart_door_lock/server/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.20.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
gunicorn>=21.0.0
